Fix Golang installation check
Since Ansible commands are executed in a non-interactive shell and our
existing Ansible Galaxy module for installing Go updates the `PATH`
environment variable globally by adding a file to /etc/profile.d, this
will not be available via the PATH var to Ansible. The 'runuser' command
ensures that profile scripts are sourced into the environment for the
command being executed, ensuring that the presence of a 'go' executable
can be determined correctly.